requirements.txt
requests>=2.28.0
schedule>=1.1.0
RPi.GPIO>=0.7.0; platform_machine=='armv7l' or platform_machine=='aarch64'
